lat,lng

+++ backend/db/models/spot.js
@@ -49,10 +49,10 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     lat: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
     },
     lng: {
-      type: DataTypes.DECIMAL
+      type: DataTypes.FLOAT
     },
     name: {
       type: DataTypes.STRING,
